pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "justllms"
version = "2.1.8"
description = "Production-focused Python library for multi-provider LLM management with unified API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "darshan harihar", email = "darshanharihar2950@gmail.com"}
]
keywords = ["llm", "ai", "openai", "anthropic", "gemini", "gateway", "proxy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.0",
    "click>=8.0.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]
providers = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "cohere>=4.0.0",
    "replicate>=0.15.0",
]
analytics = [
    "reportlab>=4.0.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
]

[project.scripts]
justllms = "justllms.cli:main"

[project.urls]
Homepage = "https://github.com/just-llms/justllms"
Documentation = "https://github.com/just-llms/justllms#readme"
Repository = "https://github.com/just-llms/justllms"
"Bug Tracker" = "https://github.com/just-llms/justllms/issues"

[tool.setuptools.packages.find]
include = ["justllms*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C90"]
ignore = ["E501", "C901"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [".venv/", "build/", "dist/"]
